convert to i32 first to avoid underflow
despite h and usedfont->h being unsigned, C allows the result to be negative

I only caught this by running the debug build in xephyr, in release I guess
something reasonable was happening automatically (probably the same as C)

+++ src/drw.rs
@@ -569,7 +569,7 @@ pub(crate) fn text(
             if utf8strlen != 0 {
                 if render != 0 {
                     ty = y
-                        + (h - (*usedfont).h) as i32 / 2
+                        + (h as i32 - (*usedfont).h as i32) / 2
                         + (*(*usedfont).xfont).ascent;
                     log::trace!("text: XftDrawStringUtf8");
                     xft::XftDrawStringUtf8(
